Master

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -43,7 +43,7 @@ return new class extends Migration
 
 
 
-            $table->integer('role')->default(1);
+            $table->unsignedBigInteger('role')->default(1);
 
             
            // $table->unsignedBigInteger('role')->default(1);
